Make Text magic comment just like all the others
This may be over-engineered. But I would like to treat this magic
comment just like all the others. On the bright side, this clears a
YAGNI path for passing arguments to Includable classes if I ever need
it.

The build was getting slow re-building everything for every page. Add
some stupid-simple caching in instance variables to speed it back up.