Move data sets into a dedicated directory
I want site and _site to be a similar shape. The leading-underscore YAML
files are not intended to have a 1:1 relationship with a generated file.

I think that's my major mental model. I want there to be a 1:1
relationship between a "source" in `site` and a "target" in `_site`.

We'll see.