Ruby is as ruby does
I like relying on `to_s` for the implicit generation of content. Now
every `eval-ruby` can be a bespoke snowflake that only gets passed data
if it absolutely needs to. Looking up data in `Build.data` is fine by
me. It might not be the most testable thing...but I haven't written any
tests for these yet, so YAGNI.