Save my data from Hey World
I like the idea of using Hey World as a simple service for building a
mailing list. If I'm going to do that, I'll want to reflect those
entries onto this website.

I don't have any immediate plans for reflecting this data. But
inspiration struck so I wrote some Ruby.