Generate an Atom feed instead of an RSS feed
I did this for changelog.xml as an experiment. Mostly because I wanted
to see what would happen if I did it.

I'm curious to see how an RSS reader handles the differences between
changes with stable ids being present. The RSS generator for RSS
wouldn't include the id attribute.