Process eval-ruby before include-in-header
I'm playing with some page specific CSS. The order of processing
matters.

My source looks like this.

```
<eval-ruby>
  <include-in-header>
    <style>/* earlier in the cascade */</style>
  </include-in-header>
</eval-ruby>
```

<include-in-header>
  <style>/* later in the cascade */</style>
</include-in-header>
```

The end result with the previous solution was:

```
<head>
  <style>/* later in the cascade */</style>
  <style>/* earlier in the cascade */</style>
</head>
```

The end result with the new solution becomes:

```
<head>
  <style>/* earlier in the cascade */</style>
  <style>/* later in the cascade */</style>
</head>
```

Which is the desired order for declarative markup and the importance of
order when defining the CSS cascade.